pyproject.toml
[project]
name = "scholarship-etl"
version = "0.1.0"
description = "etl pipelines for scholarship data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.1",
    "boto3>=1.37.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.3",
]
